correcciones

+++ app/src/main/java/com/example/appdelclima/repository/modelos/Ciudad.kt
@@ -2,8 +2,8 @@ package com.example.appdelclima.repository.modelos
 
 data class Ciudad (
     val name : String,
-    val lat : Double,
-    val lon : Double,
+    val lat : Float,
+    val lon : Float,
     val country : String,
     val state: String? = null
